+++ lib/vulkan/renpass/compo/src/slight.cpp
@@ -1,4 +1,4 @@
-#include "renderpass/compo.hpp"
+#include "mirinae/vulkan/renpass/compo/compo.hpp"
 
 #include <entt/entity/registry.hpp>
 
@@ -6,11 +6,10 @@
 #include "mirinae/cpnt/light.hpp"
 #include "mirinae/cpnt/transform.hpp"
 #include "mirinae/lightweight/task.hpp"
-
-#include "render/cmdbuf.hpp"
-#include "render/draw_set.hpp"
-#include "render/mem_cinfo.hpp"
-#include "renderpass/builder.hpp"
+#include "mirinae/vulkan/base/render/cmdbuf.hpp"
+#include "mirinae/vulkan/base/render/draw_set.hpp"
+#include "mirinae/vulkan/base/render/mem_cinfo.hpp"
+#include "mirinae/vulkan/base/renderpass/builder.hpp"
 
 
 namespace {
